Counts, CountsViewModel - Changed data type for two properties
OrderService
- Added function GetActualOrder

MainWindow.xaml
- Added new private attributes
- Added new method UpdatingCountsViewModel
- Moved view model assignation into new method

+++ Model/Counts.cs
@@ -6,8 +6,8 @@ namespace dashboard.Model
 {
     sealed class Counts
     {
-        public int Plan { get; set; }
-        public int Real { get; set; }
+        public decimal Plan { get; set; }
+        public decimal Real { get; set; }
         public DateTime TaktTime { get; set; }
         public DateTime TotalStopTime { get; set; }
         public DateTime StationsStopTime { get; set; }
